Beta 14 (#55)
* Error messages now go to stderr. Also updated CMakeLists.txt to use target_include_directories to hopefully stop json lib compile warnings

* BED format has been corrected (now includes score and strand)

* Added some new pvalue stuff.

* Shuffling now only occurs after overlap

+++ src/ultra.hpp
@@ -92,7 +92,7 @@ public:
 
   void CorrectOverlap(int maxReadID);
 
-  float Log2PvalForScore(float score, float period) const;
+  double Log2PvalForScore(float score, float period) const;
 
   SequenceWindow *GetSequenceWindow(SequenceWindow *seq);
